Fixed warning.
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@32861 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/opal/transcoders.cxx
@@ -110,7 +110,7 @@ PBoolean OpalTranscoder::ExecuteCommand(const OpalMediaCommand & command)
     const OpalMediaFlowControl * flow = dynamic_cast<const OpalMediaFlowControl *>(&command);
     if (flow != NULL) {
       unsigned bitRate = std::min(flow->GetMaxBitRate(), outputMediaFormat.GetMaxBandwidth());
-      if (outputMediaFormat.GetOptionInteger(OpalMediaFormat::TargetBitRateOption()) != bitRate) {
+      if ((unsigned)outputMediaFormat.GetOptionInteger(OpalMediaFormat::TargetBitRateOption()) != bitRate) {
         outputMediaFormat.SetOptionInteger(OpalMediaFormat::TargetBitRateOption(), bitRate);
         UpdateMediaFormats(OpalMediaFormat(), outputMediaFormat);
       }
